2nd Commit

+++ app/Http/Controllers/Country/CountryController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 use App\Models\CountryModel;
+use Validator;
 
 class CountryController extends Controller
 {
@@ -18,14 +19,22 @@ class CountryController extends Controller
     public function countryByID($id){
         $country = CountryModel::find($id);
         if(is_null($country)){
-            return response()->json('Record Not Found!',404);
+            return response()->json(["message" => "Record not found!"],404);
         }
         return response()->json($country,200);
-    
+        
     }
-
+    
     // CREATE
     public function countrySave(Request $request){
+        $rules = [
+            'name' => 'required|min:3',
+            'iso' => 'required|min:2|max:2|unique'
+        ];
+        $validator = Validator::make($request->all(),$rules);
+        if($validator->fails()){
+            return response()->json($validator->errors(),400);
+        }
         $country = CountryModel::create($request->all());
         return response()->json($country, 201);
     }
@@ -33,12 +42,19 @@ class CountryController extends Controller
     // UPDATE
     public function countryUpdate(Request $request, $id){
         $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()->json(["message" => "Record not found!"],404);
+        }
         $country->update($request->all());
         return response()->json($country,200);
     }
 
     // DELETE
-    public function countryDelete(Request $request, CountryModel $country){
+    public function countryDelete(Request $request, $id){
+        $country = CountryModel::delete($id);
+        if(is_null($country)){
+            return response()->json(["message" => "Record not found!"],404);
+        }
         $country->delete();
         return response()->json(null,204);
     }
